+++ src/components/layout/PageWrapper/PageWrapper.style.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  position: relative;
+  width: 100vw;
+
+  padding: 0 160px 0 160px;
+  box-sizing: border-box;
+  @media screen and (max-width: 1440px) {
+    padding: 0 40px 0 40px;
+  }
+`;
